refactor: deploy command
