+++ .env.example
@@ -7,3 +7,4 @@ SMTP_PORT=
 SMTP_USER=
 SMTP_PASS=
 SMTP_SECURE=
+BOT_OWNER_ID=
