+++ src/discord/commands/index.ts
@@ -1,5 +1,4 @@
 import Ping from './ping';
-import Deploy from './deploy';
 import Rooms from './rooms';
 import Book from './book';
 import Manage from './manage';
@@ -7,4 +6,4 @@ import View from './view';
 import Report from './report';
 import Delete from './delete';
 
-export default [Deploy, Rooms, Book, Manage, View, Delete, Report, Ping];
+export default [Rooms, Book, Manage, View, Delete, Report, Ping];
